backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1 # You'll likely need Uvicorn to run your FastAPI app
python-dotenv==1.0.0
PyMuPDF==1.24.5 # This is the 'fitz' library
python-docx==1.1.2
requests==2.32.3